+++ includes/third-party/class-cocart-third-party.php
@@ -25,6 +25,7 @@ if ( ! class_exists( 'CoCart_Third_Party' ) ) {
 		 */
 		public function __construct() {
 			self::include_hosts();
+			self::include_plugins();
 		}
 
 		/**
@@ -35,6 +36,16 @@ if ( ! class_exists( 'CoCart_Third_Party' ) ) {
 		 */
 		public function include_hosts() {
 			include_once COCART_ABSPATH . 'includes/third-party/hosting/pantheon/class-cocart-host-pantheon.php'; // Pantheon.io
+		}
+
+		/**
+		 * Load support for third-party plugins.
+		 *
+		 * @access public
+		 * @return string
+		 */
+		public function include_plugins() {
+			include_once COCART_ABSPATH . 'includes/third-party/plugin/taxjar/class-cocart-plugin-taxjar.php'; // TaxJar
 		}
 
 	} // END class.